fpi/requirements.txt
fastapi
pydantic
python-dotenv
groq
uvicorn